requirements.txt
# Add your dependencies here
numpy
pandas
scikit-learn
geopandas
rasterio
twilio>=8.10.0
python-dotenv>=1.0.0